Extract data transfer objects into own file
This allows us resuse the types and conversion functions for different apis (sse or http)
